update go.mod for clusteragent endpoint
