+++ comp/forwarder/defaultforwarder/go.mod
@@ -31,6 +31,7 @@ replace (
 	github.com/DataDog/datadog-agent/pkg/trace => ../../../pkg/trace
 	github.com/DataDog/datadog-agent/pkg/util/backoff => ../../../pkg/util/backoff
 	github.com/DataDog/datadog-agent/pkg/util/common => ../../../pkg/util/common
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint => ../../../pkg/util/clusteragent/endpoint
 	github.com/DataDog/datadog-agent/pkg/util/defaultpaths => ../../../pkg/util/defaultpaths
 	github.com/DataDog/datadog-agent/pkg/util/executable => ../../../pkg/util/executable
 	github.com/DataDog/datadog-agent/pkg/util/filesystem => ../../../pkg/util/filesystem
@@ -64,6 +65,7 @@ require (
 	github.com/DataDog/datadog-agent/pkg/telemetry v0.56.0-rc.3
 	github.com/DataDog/datadog-agent/pkg/util/backoff v0.56.0-rc.3
 	github.com/DataDog/datadog-agent/pkg/util/common v0.56.0-rc.3
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint v0.0.1
 	github.com/DataDog/datadog-agent/pkg/util/filesystem v0.57.1
 	github.com/DataDog/datadog-agent/pkg/util/fxutil v0.56.0-rc.3
 	github.com/DataDog/datadog-agent/pkg/util/http v0.56.0-rc.3
