+++ pkg/serializer/go.mod
@@ -40,6 +40,7 @@ replace (
 	github.com/DataDog/datadog-agent/pkg/trace => ../trace
 	github.com/DataDog/datadog-agent/pkg/util/backoff => ../util/backoff/
 	github.com/DataDog/datadog-agent/pkg/util/buf => ../util/buf/
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint => ../util/clusteragent/endpoint
 	github.com/DataDog/datadog-agent/pkg/util/common => ../util/common
 	github.com/DataDog/datadog-agent/pkg/util/defaultpaths => ../../pkg/util/defaultpaths
 	github.com/DataDog/datadog-agent/pkg/util/executable => ../util/executable/
@@ -106,6 +107,7 @@ require (
 	github.com/DataDog/datadog-agent/pkg/status/health v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/backoff v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/buf v0.56.0-rc.3 // indirect
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint v0.0.1 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/common v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/executable v0.57.1 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/filesystem v0.57.1 // indirect
