+++ test/otel/go.mod
@@ -68,6 +68,7 @@ replace (
 	github.com/DataDog/datadog-agent/pkg/util/backoff => ./../../pkg/util/backoff
 	github.com/DataDog/datadog-agent/pkg/util/buf => ../../pkg/util/buf
 	github.com/DataDog/datadog-agent/pkg/util/cgroups => ./../../pkg/util/cgroups
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint => ../../pkg/util/clusteragent/endpoint
 	github.com/DataDog/datadog-agent/pkg/util/common => ../../pkg/util/common
 	github.com/DataDog/datadog-agent/pkg/util/defaultpaths => ./../../pkg/util/defaultpaths
 	github.com/DataDog/datadog-agent/pkg/util/executable => ./../../pkg/util/executable
@@ -156,6 +157,7 @@ require (
 	github.com/DataDog/datadog-agent/pkg/util/backoff v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/buf v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/cgroups v0.56.0-rc.3 // indirect
+	github.com/DataDog/datadog-agent/pkg/util/clusteragent/endpoint v0.0.1 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/common v0.56.0-rc.3 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/executable v0.57.1 // indirect
 	github.com/DataDog/datadog-agent/pkg/util/filesystem v0.57.1 // indirect
